Extend with O(log n) spare ID generation
Imagine a BTree used to store a set of entitites keyed by ID.  When a
new entity is created, it needs a distinct ID.  The tree could be used
to check if the given ID was already taken; if it was, attempt the next
one.  This algorithm might perform numerous lookups, which is suboptimal.

By extending tree nodes with additional metadata, we are able to answer
"give me the first unused ID greater or equal to the given value" with
log(n) complexity. The complexity of the base insert/delete/get
operations is unchanged.

This commit adds FirstWithoutSuccessor method to BTree:

func (t *BTree) FirstWithoutSuccessor(greaterOrEqualTo Item) Item

Intended usage:

if k := t.FirstWithoutSuccessor(nextID); k == nil {
    id = nextID
} else {
    id = k + 1
}